+++ crates/engine/src/lib.rs
@@ -2,6 +2,7 @@ mod timing;
 pub mod cell;
 pub mod cell_id;
 pub mod cond_format;
+pub mod custom_fns;
 pub mod dep_graph;
 pub mod events;
 pub mod filter;